implements znode payment from getblocktemplate information
make sure znode payment is send to the address given by getblocktemplate. take also care of some weird even when the znode field is empty (test).
This version should work on both test mainnet and standard mtp (however founder addresses are the mainnet one, this is possible due to the absence of verification on testnet)